better lanyard
