+++ app/(home)/lanyard/index.ts
@@ -1,0 +1,2 @@
+export { default as Lanyard } from "./lanyard";
+export { default as LanyardSkeleton } from "./skeleton";
